+++ packages/stacked_tools/lib/src/templates/compiled_templates.dart
@@ -1,6 +1,7 @@
 /// NOTE: This is generated code from the compileTemplates command. Do not modify by hand
 ///       This file should be checked into source control.
 
+
 // -------- HomeViewmodelTest Template Data ----------
 
 const String kAppTemplateHomeViewmodelTestPath =
@@ -52,9 +53,11 @@ void main() {
 
 // --------------------------------------------------
 
+
 // -------- TestHelpers Template Data ----------
 
-const String kAppTemplateTestHelpersPath = 'test/helpers/test_helpers.dart.stk';
+const String kAppTemplateTestHelpersPath =
+    'test/helpers/test_helpers.dart.stk';
 
 const String kAppTemplateTestHelpersContent = '''
 import 'package:mockito/annotations.dart';
@@ -140,9 +143,11 @@ void _removeRegistrationIfExists<T extends Object>() {
 
 // --------------------------------------------------
 
+
 // -------- StackedConfigJsonStk Template Data ----------
 
-const String kAppTemplateStackedConfigJsonStkPath = 'stacked.config.json.stk';
+const String kAppTemplateStackedConfigJsonStkPath =
+    'stacked.config.json.stk';
 
 const String kAppTemplateStackedConfigJsonStkContent = '''
 {
@@ -157,9 +162,11 @@ const String kAppTemplateStackedConfigJsonStkContent = '''
 
 // --------------------------------------------------
 
+
 // -------- READMEMdStk Template Data ----------
 
-const String kAppTemplateREADMEMdStkPath = 'README.md.stk';
+const String kAppTemplateREADMEMdStkPath =
+    'README.md.stk';
 
 const String kAppTemplateREADMEMdStkContent = '''
 # stacked_app
@@ -183,9 +190,11 @@ samples, guidance on mobile development, and a full API reference.
 
 // --------------------------------------------------
 
+
 // -------- Main Template Data ----------
 
-const String kAppTemplateMainPath = 'lib/main.dart.stk';
+const String kAppTemplateMainPath =
+    'lib/main.dart.stk';
 
 const String kAppTemplateMainContent = '''
 import 'package:flutter/material.dart';
@@ -230,6 +239,7 @@ class MyApp extends StatelessWidget {
 
 // --------------------------------------------------
 
+
 // -------- SetupDialogUi Template Data ----------
 
 const String kAppTemplateSetupDialogUiPath =
@@ -255,6 +265,7 @@ void setupDialogUi() {
 ''';
 
 // --------------------------------------------------
+
 
 // -------- SetupBottomSheetUi Template Data ----------
 
@@ -286,9 +297,11 @@ void setupBottomSheetUi() {
 
 // --------------------------------------------------
 
+
 // -------- UiHelpers Template Data ----------
 
-const String kAppTemplateUiHelpersPath = 'lib/ui/common/ui_helpers.dart.stk';
+const String kAppTemplateUiHelpersPath =
+    'lib/ui/common/ui_helpers.dart.stk';
 
 const String kAppTemplateUiHelpersContent = '''
 import 'dart:math';
@@ -374,9 +387,11 @@ double getResponsiveFontSize(BuildContext context,
 
 // --------------------------------------------------
 
+
 // -------- AppStrings Template Data ----------
 
-const String kAppTemplateAppStringsPath = 'lib/ui/common/app_strings.dart.stk';
+const String kAppTemplateAppStringsPath =
+    'lib/ui/common/app_strings.dart.stk';
 
 const String kAppTemplateAppStringsContent = '''
 const String ksHomeBottomSheetTitle = 'Build Great Apps!';
@@ -387,9 +402,11 @@ const String ksHomeBottomSheetDescription =
 
 // --------------------------------------------------
 
+
 // -------- AppColors Template Data ----------
 
-const String kAppTemplateAppColorsPath = 'lib/ui/common/app_colors.dart.stk';
+const String kAppTemplateAppColorsPath =
+    'lib/ui/common/app_colors.dart.stk';
 
 const String kAppTemplateAppColorsContent = '''
 import 'package:flutter/material.dart';
@@ -405,6 +422,7 @@ const Color kcBackgroundColor = kcDarkGreyColor;
 ''';
 
 // --------------------------------------------------
+
 
 // -------- NoticeSheet Template Data ----------
 
@@ -463,6 +481,7 @@ class NoticeSheet extends StatelessWidget {
 
 // --------------------------------------------------
 
+
 // -------- NoticeSheetViewmodel Template Data ----------
 
 const String kAppTemplateNoticeSheetViewmodelPath =
@@ -473,6 +492,7 @@ const String kAppTemplateNoticeSheetViewmodelContent = '''
 ''';
 
 // --------------------------------------------------
+
 
 // -------- InfoAlertDialog Template Data ----------
 
@@ -582,6 +602,7 @@ class InfoAlertDialog extends StatelessWidget {
 
 // --------------------------------------------------
 
+
 // -------- InfoAlertViewmodel Template Data ----------
 
 const String kAppTemplateInfoAlertViewmodelPath =
@@ -595,9 +616,11 @@ const String kAppTemplateInfoAlertViewmodelContent = '''
 
 // --------------------------------------------------
 
+
 // -------- HomeView Template Data ----------
 
-const String kAppTemplateHomeViewPath = 'lib/ui/views/home/home_view.dart.stk';
+const String kAppTemplateHomeViewPath =
+    'lib/ui/views/home/home_view.dart.stk';
 
 const String kAppTemplateHomeViewContent = '''
 import 'package:flutter/material.dart';
@@ -683,6 +706,7 @@ class HomeView extends StatelessWidget {
 
 // --------------------------------------------------
 
+
 // -------- HomeViewmodel Template Data ----------
 
 const String kAppTemplateHomeViewmodelPath =
@@ -729,6 +753,7 @@ class HomeViewModel extends BaseViewModel {
 
 // --------------------------------------------------
 
+
 // -------- StartupViewmodel Template Data ----------
 
 const String kAppTemplateStartupViewmodelPath =
@@ -757,6 +782,7 @@ class StartupViewModel extends BaseViewModel {
 ''';
 
 // --------------------------------------------------
+
 
 // -------- StartupView Template Data ----------
 
@@ -824,9 +850,11 @@ class StartupView extends StatelessWidget {
 
 // --------------------------------------------------
 
+
 // -------- App Template Data ----------
 
-const String kAppTemplateAppPath = 'lib/app/app.dart.stk';
+const String kAppTemplateAppPath =
+    'lib/app/app.dart.stk';
 
 const String kAppTemplateAppContent = '''
 import 'package:stacked/stacked_annotations.dart';
@@ -851,9 +879,11 @@ class App {}
 
 // --------------------------------------------------
 
+
 // -------- DialogType Template Data ----------
 
-const String kAppTemplateDialogTypePath = 'lib/enums/dialog_type.dart.stk';
+const String kAppTemplateDialogTypePath =
+    'lib/enums/dialog_type.dart.stk';
 
 const String kAppTemplateDialogTypeContent = '''
 enum DialogType {
@@ -863,6 +893,7 @@ enum DialogType {
 ''';
 
 // --------------------------------------------------
+
 
 // -------- BottomSheetType Template Data ----------
 
@@ -878,9 +909,11 @@ enum BottomSheetType {
 
 // --------------------------------------------------
 
+
 // -------- PubspecYamlStk Template Data ----------
 
-const String kAppTemplatePubspecYamlStkPath = 'pubspec.yaml.stk';
+const String kAppTemplatePubspecYamlStkPath =
+    'pubspec.yaml.stk';
 
 const String kAppTemplatePubspecYamlStkContent = '''
 name: {{packageName}}
@@ -984,6 +1017,7 @@ flutter:
 
 // --------------------------------------------------
 
+
 // -------- GenericViewmodelTest Template Data ----------
 
 const String kViewTemplateGenericViewmodelTestPath =
@@ -1006,6 +1040,7 @@ void main() {
 
 // --------------------------------------------------
 
+
 // -------- GenericViewmodel Template Data ----------
 
 const String kViewTemplateGenericViewmodelPath =
@@ -1019,6 +1054,7 @@ class {{viewModelName}} extends BaseViewModel {
 ''';
 
 // --------------------------------------------------
+
 
 // -------- GenericView Template Data ----------
 
@@ -1051,6 +1087,7 @@ class {{viewName}} extends StatelessWidget {
 
 // --------------------------------------------------
 
+
 // -------- GenericServiceTest Template Data ----------
 
 const String kServiceTemplateGenericServiceTestPath =
@@ -1073,6 +1110,7 @@ void main() {
 
 // --------------------------------------------------
 
+
 // -------- GenericService Template Data ----------
 
 const String kServiceTemplateGenericServicePath =
@@ -1085,3 +1123,4 @@ class {{serviceName}}Service {
 ''';
 
 // --------------------------------------------------
+